write tests to uncover bug; still need to decide how to fix it.
This function is broken:
        fn get(&self) -> Option<&ManaAbilitiesSet> {
            if self.fully_advanced() {
                None
            } else {
                Some(&self.current_set)
            }
        }
This is because, by design, once the iter is at the last element, it is fully advanced. But we want to be able to get the last element once!

Maybe see what the streaming iterator package does? A one off boolean seems kind of inelegant.

But a bool called "past_end" that we check in get and set in get if we get while fully_advanced, or if we advance while already fully_advanced seems like it would work?